refactor: 🎨 First idea of utils refactor
