+++ src/bot/utils/universal_options/interactions.ts
@@ -1,0 +1,31 @@
+import { InteractionReplyOptions } from 'discord.js';
+import { info, problem, validate } from '../embeds';
+
+type UniversalInteractionOptions = Omit<InteractionReplyOptions, 'flags'>;
+
+export function interactionInfo(message: string, ephemeral = true): UniversalInteractionOptions {
+  return {
+    content: '',
+    embeds: [info(message)],
+    components: [],
+    ephemeral
+  };
+}
+
+export function interactionProblem(message: string, ephemeral = true): UniversalInteractionOptions {
+  return {
+    content: '',
+    embeds: [problem(message)],
+    components: [],
+    ephemeral
+  };
+}
+
+export function interactionValidate(message: string, ephemeral = true): UniversalInteractionOptions {
+  return {
+    content: '',
+    embeds: [validate(message)],
+    components: [],
+    ephemeral
+  };
+}
