+++ src/bot/utils/universal_options/messages.ts
@@ -1,0 +1,28 @@
+import { MessageOptions } from 'discord.js';
+import { info, problem, validate } from '../embeds';
+
+type UniversalMessageOptions = Omit<MessageOptions, 'flags'>;
+
+export function messageInfo(message: string): UniversalMessageOptions {
+  return {
+    content: '',
+    embeds: [info(message)],
+    components: []
+  };
+}
+
+export function messageProblem(message: string): UniversalMessageOptions {
+  return {
+    content: '',
+    embeds: [problem(message)],
+    components: []
+  };
+}
+
+export function messageValidate(message: string): UniversalMessageOptions {
+  return {
+    content: '',
+    embeds: [validate(message)],
+    components: []
+  };
+}
