+++ src/bot/utils/utils.ts
@@ -1,0 +1,35 @@
+import { TextChannel, GuildMember, Message } from 'discord.js';
+import { diffWords } from 'diff';
+
+import { godfatherRoleId } from '../constants';
+
+export function isEmbedable(channel: TextChannel): boolean {
+  const permissions = channel.permissionsFor(channel.guild.members.me!);
+  return permissions?.has(['ViewChannel', 'SendMessages', 'EmbedLinks']);
+}
+
+export function isGodfather(member: GuildMember): boolean {
+  return member.roles.cache.has(godfatherRoleId);
+}
+
+export function messageLink(guildId: string, channelId: string, messageId: string): string {
+  return `https://discord.com/channels/${guildId}/${channelId}/${messageId}`;
+}
+
+export function tDelete(timeout = 6000): (message: Message) => NodeJS.Timeout {
+  return (message: Message) => setTimeout(() => message.deletable && message.delete().catch(() => null), timeout);
+}
+
+export function showPositiveDiffs(oldValue: string, newValue: string): string {
+  return diffWords(oldValue, newValue)
+    .filter((part) => !part.removed)
+    .map((part) => `${part.added ? '`' : ''}${part.value}${part.added ? '`' : ''}`)
+    .join('');
+}
+
+export function showNegativeDiffs(oldValue: string, newValue: string): string {
+  return diffWords(oldValue, newValue)
+    .filter((part) => !part.added)
+    .map((part) => `${part.removed ? '~~`' : ''}${part.value}${part.removed ? '`~~' : ''}`)
+    .join('');
+}
